Move default definitions
Moved the definitions of default command-line parameter values out of
`apidump` and into `apispec`, to avoid having `apispec` depend on
`apidump`. This allows client code to import `apispec` without having to
also pull in pcap.
