+++ cmd/internal/apidump/apidump.go
@@ -8,6 +8,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/akitasoftware/akita-cli/apidump"
+	"github.com/akitasoftware/akita-cli/apispec"
 	"github.com/akitasoftware/akita-cli/cmd/internal/cmderr"
 	"github.com/akitasoftware/akita-cli/cmd/internal/pluginloader"
 	"github.com/akitasoftware/akita-cli/location"
@@ -114,13 +115,13 @@ var Cmd = &cobra.Command{
 				}
 			} else {
 				if outFlag.AkitaURI.ObjectName == "" {
-					traceRotateInterval = apidump.DefaultTraceRotateInterval
+					traceRotateInterval = apispec.DefaultTraceRotateInterval
 				}
 			}
 		}
 
 		if deploymentFlag == "" {
-			deploymentFlag = apidump.DefaultDeployment
+			deploymentFlag = apispec.DefaultDeployment
 			if os.Getenv("AKITA_DEPLOYMENT") != "" {
 				deploymentFlag = os.Getenv("AKITA_DEPLOYMENT")
 			}
@@ -230,7 +231,7 @@ func init() {
 	Cmd.Flags().Float64Var(
 		&rateLimitFlag,
 		"rate-limit",
-		apidump.DefaultRateLimit,
+		apispec.DefaultRateLimit,
 		"Number of requests per minute to capture.",
 	)
 
@@ -319,14 +320,14 @@ func init() {
 	Cmd.Flags().IntVar(
 		&statsLogDelay,
 		"stats-log-delay",
-		apidump.DefaultStatsLogDelay_seconds,
+		apispec.DefaultStatsLogDelay_seconds,
 		"Print packet capture statistics after N seconds.",
 	)
 
 	Cmd.Flags().IntVar(
 		&telemetryInterval,
 		"telemetry-interval",
-		apidump.DefaultTelemetryInterval_seconds,
+		apispec.DefaultTelemetryInterval_seconds,
 		"Upload client telemetry every N seconds.",
 	)
 	Cmd.Flags().MarkHidden("telemetry-interval")
@@ -334,7 +335,7 @@ func init() {
 	Cmd.Flags().BoolVar(
 		&collectTCPAndTLSReports,
 		"report-tcp-and-tls",
-		apidump.DefaultCollectTCPAndTLSReports,
+		apispec.DefaultCollectTCPAndTLSReports,
 		"Collect TCP and TLS reports.",
 	)
 	Cmd.Flags().MarkHidden("report-tcp-and-tls")
@@ -342,7 +343,7 @@ func init() {
 	Cmd.Flags().BoolVar(
 		&parseTLSHandshakes,
 		"parse-tls-handshakes",
-		apidump.DefaultParseTLSHandshakes,
+		apispec.DefaultParseTLSHandshakes,
 		"Parse TLS handshake packets.",
 	)
 	Cmd.Flags().MarkHidden("parse-tls-handshakes")
@@ -350,7 +351,7 @@ func init() {
 	Cmd.Flags().IntVar(
 		&maxWitnessSize_bytes,
 		"max-witness-size-bytes",
-		apidump.DefaultMaxWitnessSize_bytes,
+		apispec.DefaultMaxWitnessSize_bytes,
 		"Don't send witnesses larger than this.",
 	)
 	Cmd.Flags().MarkHidden("max-witness-size-bytes")
