+++ daemon/internal/cloud_client/cloud_client.go
@@ -3,7 +3,6 @@ package cloud_client
 import (
 	"fmt"
 
-	"github.com/akitasoftware/akita-cli/apidump"
 	"github.com/akitasoftware/akita-cli/apispec"
 	"github.com/akitasoftware/akita-cli/plugin"
 	"github.com/akitasoftware/akita-cli/printer"
@@ -160,7 +159,7 @@ func collectTraces(traceEventChannel <-chan *TraceEvent, learnClient rest.LearnC
 		loggingOptions.TraceID,
 		learnClient,
 		// TODO Make this configurable.
-		optionals.Some(apidump.DefaultMaxWitnessSize_bytes),
+		optionals.Some(apispec.DefaultMaxWitnessSize_bytes),
 		plugins,
 	)
 	collector = &trace.PacketCountCollector{
